+++ cassandra-test-image/src/main/docker/ecc-entrypoint.sh
@@ -127,7 +127,7 @@ if [ "$(echo -n "$JOLOKIA" | xargs)" = "true" ]; then
     echo "Attempting to download Jolokia jar..."
     curl -L -o "$JOLOKIA_JAR" "https://search.maven.org/remotecontent?filepath=org/jolokia/jolokia-agent-jvm/2.1.1/jolokia-agent-jvm-2.1.1-javaagent.jar"
 
-    JOLOKIA_OPTS="--add-opens=java.base/jdk.internal.ref=ALL-UNNAMED -javaagent:$JOLOKIA_JAR=port=8778,host=0.0.0.0,useSsl=false"
+    JOLOKIA_OPTS="--add-opens=java.base/jdk.internal.ref=ALL-UNNAMED -javaagent:$JOLOKIA_JAR=port=8778,host=*,useSsl=false"
 
     if [ -n "$JVM_EXTRA_OPTS" ]; then
         export JVM_EXTRA_OPTS="$JVM_EXTRA_OPTS $JOLOKIA_OPTS"