+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Version 1.0.0 (Not yet Released)
 
+* Create Integration Tests Using the New Jolokia Functionality - Issue #845
+* ClassCastException when using Jolokia for JMX metrics retrieval - Issue #1126
 * ecc.sh in evolved ecchronos uses old SpringBooter package reference - Issue #1123
 * ecctool with node and keyspace gives an error -Issue #1055
 * REST Interface to provide list of nodes - Issue #1045