+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Changes
 
+## Version 6.1.4
+
 ## Version 6.1.3
 
 * Rework the status commands of the ecctool for more details - Issue #946